requirements.txt
pandas==2.2.2
numpy==1.23.5
nbformat==5.10.4
plotly==5.24.0
pyod==2.0.2 
scikit-learn==1.5.1 
scipy==1.5.1 
coverage==7.3.1
tox==4.8.0,<=4.11.3
xgboost==1.6.2
pyarrow==17.0.0
lightgbm==4.5.0
skim==1.2
pyod==2.0.2
yellowbrick==1.5
seaborn==0.13.2
hpsklearn==0.1.0
hyperopt==0.2.7
hyperopt>=0.2.6,<=0.2.7

polars==1.6.0
skimpy==0.0.15
tpot==0.12.2
ray==2.36.0
torch==2.4.1
scikit-optimize==0.10.2
catboost==1.2.7
graphviz==0.20.3
bayesian-optimization==1.5.1
statsmodels==0.14.3
mlxtend==0.23.1
eli5==0.13.0
tensorflow==2.17.0
tensorflow_decision_forests==1.10.0
seaborn==0.13.2